drf_sp_hub/requirements.txt
Django==2.0
django-rest-framework==0.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.3
psycopg2==2.7.3.2
pytz==2017.3
